Cleanup, re-organized utils package
