+++ packages/text-annotator/src/utils/index.ts
@@ -1,8 +1,13 @@
+export * from './debounce';
+export * from './getAnnotatableFragment';
 export * from './getClientRectsPonyfill';
-export * from './mergeClientRects';
-export * from './trimRange';
-export * from './getAnnotatableRanges';
 export * from './getQuoteContext';
+export * from './isRevived';
+export * from './mergeClientRects';
 export * from './rangeToSelector';
 export * from './reviveAnnotation';
-export * from './debounce';
+export * from './reviveSelector';
+export * from './reviveTarget';
+export * from './splitAnnotatableRanges';
+export * from './trimRange';
+
