+++ packages/text-annotator/src/utils/isRevived.ts
@@ -1,0 +1,4 @@
+import type { TextSelector } from '../model';
+
+export const isRevived = (selector: TextSelector[]) =>
+  selector.every(s => s.range instanceof Range && !s.range.collapsed);
